# Fix metric confusion when the metric name is contained in the tags