--- a/pkg/aggregator/ckey/key_test.go
+++ b/pkg/aggregator/ckey/key_test.go
@@ -27,7 +27,7 @@ func TestGenerateReproductible(t *testing.T) {
 	generator := NewKeyGenerator()
 
 	firstKey := generator.Generate(name, hostname, tags)
-	assert.Equal(t, ContextKey(0x932f9848b0fb0802), firstKey)
+	assert.Equal(t, ContextKey(0x1e923504c1aad3ad), firstKey)
 
 	for n := 0; n < 10; n++ {
 		t.Run(fmt.Sprintf("iteration %d:", n), func(t *testing.T) {
@@ -38,7 +38,7 @@ func TestGenerateReproductible(t *testing.T) {
 
 	otherKey := generator.Generate("othername", hostname, tags)
 	assert.NotEqual(t, firstKey, otherKey)
-	assert.Equal(t, ContextKey(0xb059e8f73b4b7ae0), otherKey)
+	assert.Equal(t, ContextKey(0xd298ae9740130f30), otherKey)
 }
 
 func TestGenerateReproductible2(t *testing.T) {
@@ -50,7 +50,7 @@ func TestGenerateReproductible2(t *testing.T) {
 	generator := NewKeyGenerator()
 
 	firstKey, tagsKey1, tagsKey2 := generator.GenerateWithTags2(name, hostname, tags1, tags2)
-	assert.Equal(t, ContextKey(0x932f9848b0fb0802), firstKey)
+	assert.Equal(t, ContextKey(0x1e923504c1aad3ad), firstKey)
 	assert.Equal(t, TagsKey(0x437b13a371a1c7d3), tagsKey1)
 	assert.Equal(t, TagsKey(0), tagsKey2)
 
@@ -65,9 +65,30 @@ func TestGenerateReproductible2(t *testing.T) {
 
 	otherKey, otherTagsKey1, otherTagsKey2 := generator.GenerateWithTags2("othername", hostname, tags1, tags2)
 	assert.NotEqual(t, firstKey, otherKey)
-	assert.Equal(t, ContextKey(0xb059e8f73b4b7ae0), otherKey)
+	assert.Equal(t, ContextKey(0xd298ae9740130f30), otherKey)
 	assert.Equal(t, tagsKey1, otherTagsKey1)
 	assert.Equal(t, tagsKey2, otherTagsKey2)
+}
+
+func TestMetricTagOverlap(t *testing.T) {
+	g := NewKeyGenerator()
+
+	empty := tagset.NewHashingTagsAccumulator()
+	h1, _, _ := g.GenerateWithTags2("metric1", "hostname",
+		tagset.NewHashingTagsAccumulatorWithTags([]string{"metric1", "t1", "t2"}), empty)
+	h2, _, _ := g.GenerateWithTags2("metric2", "hostname",
+		tagset.NewHashingTagsAccumulatorWithTags([]string{"metric2", "t1", "t2"}), empty)
+
+	assert.NotEqual(t, h1, h2)
+}
+
+func TestMetricHostnameSplit(t *testing.T) {
+	g := NewKeyGenerator()
+	empty := tagset.NewHashingTagsAccumulator()
+	h1, _, _ := g.GenerateWithTags2("metric", "hostname", empty, empty)
+	h2, _, _ := g.GenerateWithTags2("metrichost", "name", empty, empty)
+
+	assert.NotEqual(t, h1, h2)
 }
 
 func TestCompare(t *testing.T) {