### What does this PR do?

Previously, when the metric name occurred in the tags list, the two
hashes would cancel out when we combine them with XOR, causing the
same context key to be returned for different metrics.

This patch replaces XOR hash combination with a murmur3 hash of metric
name, hostname and the bytes of the tags hash.


### Motivation

Avoid metric confusion when metric name is present in the tags.

### Possible Drawbacks / Trade-offs

### Describe how to test/QA your changes

```
docker run --rm --name qa \
-e DD_LOG_LEVEL=debug -e DD_LOG_PAYLOADS=true -e DD_DOGSTATSD_SOCKET=/tmp/dsd.sock \
-e DD_API_KEY=none -v /tmp:/tmp \
datadog/agent:7.36.0-rc.1
```

Send two or more metrics that both have the metric name in the tags:
```
echo -e 'metric.1:11111|g|#metric.1\nmetric.2:22222|g|#metric.2' | nc -uU /tmp/dsd.sock
```

Verify that both metrics are present in the outgoing payload:

```
docker logs qa | grep metric\\\.
```

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [x] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [x] The appropriate `team/..` label has been applied, if known.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [x] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.