@djmitche 2022-04-07
From before this is merged to 7.36.0-rc.1:
```
⸩ benchstat before.txt after.txt
name                         old time/op  new time/op  delta
KeyGeneration/1-tags-16      25.2ns ± 0%  29.8ns ± 0%  +18.36%  (p=0.008 n=5+5)
KeyGeneration/2-tags-16      27.7ns ± 0%  30.9ns ± 0%  +11.62%  (p=0.008 n=5+5)
KeyGeneration/4-tags-16      35.5ns ± 0%  37.3ns ± 0%   +5.28%  (p=0.008 n=5+5)
KeyGeneration/8-tags-16      48.6ns ± 0%  50.7ns ± 1%   +4.46%  (p=0.008 n=5+5)
KeyGeneration/16-tags-16     75.7ns ± 0%  78.4ns ± 0%   +3.56%  (p=0.008 n=5+5)
KeyGeneration/32-tags-16      121ns ± 0%   124ns ± 2%   +2.58%  (p=0.016 n=4+5)
KeyGeneration/64-tags-16      223ns ± 0%   225ns ± 1%   +0.77%  (p=0.008 n=5+5)
KeyGeneration/128-tags-16     445ns ± 2%   441ns ± 1%     ~     (p=0.690 n=5+5)
KeyGeneration/256-tags-16     832ns ± 1%   831ns ± 0%     ~     (p=0.690 n=5+5)
KeyGeneration/512-tags-16    12.9µs ± 0%  12.9µs ± 0%   +0.32%  (p=0.016 n=5+5)
KeyGeneration/1024-tags-16   68.7µs ± 1%  68.2µs ± 0%     ~     (p=0.151 n=5+5)
KeyGeneration/2048-tags-16    164µs ± 1%   163µs ± 2%     ~     (p=0.151 n=5+5)
KeyGeneration2/1-tags-16     25.2ns ± 0%  29.4ns ± 0%  +16.84%  (p=0.008 n=5+5)
KeyGeneration2/2-tags-16     27.5ns ± 0%  30.7ns ± 0%  +11.71%  (p=0.016 n=5+4)
KeyGeneration2/4-tags-16     34.9ns ± 0%  37.4ns ± 0%   +6.97%  (p=0.008 n=5+5)
KeyGeneration2/8-tags-16     48.4ns ± 0%  50.8ns ± 0%   +5.00%  (p=0.008 n=5+5)
KeyGeneration2/16-tags-16    76.0ns ± 0%  78.4ns ± 1%   +3.23%  (p=0.008 n=5+5)
KeyGeneration2/32-tags-16     122ns ± 4%   124ns ± 0%     ~     (p=0.190 n=5+4)
KeyGeneration2/64-tags-16     223ns ± 1%   226ns ± 1%   +1.06%  (p=0.008 n=5+5)
KeyGeneration2/128-tags-16    441ns ± 1%   441ns ± 1%     ~     (p=0.841 n=5+5)
KeyGeneration2/256-tags-16    827ns ± 1%   830ns ± 1%     ~     (p=0.151 n=5+5)
KeyGeneration2/512-tags-16   12.9µs ± 0%  12.9µs ± 0%     ~     (p=0.190 n=5+5)
KeyGeneration2/1024-tags-16  68.5µs ± 1%  68.2µs ± 0%     ~     (p=0.095 n=5+5)
KeyGeneration2/2048-tags-16   163µs ± 1%   163µs ± 0%     ~     (p=0.095 n=5+5)
InsertionSort-16             9.59µs ± 0%  9.59µs ± 0%     ~     (p=1.000 n=5+5)
StdlibSort-16                9.92µs ± 0%  9.93µs ± 0%     ~     (p=0.175 n=5+5)
```